requirements.txt
ipython
black
mypy
pillow
types-pillow
tqdm
types-tqdm
